requirements.txt
# Core PyTorch (install with appropriate CUDA version)
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Scientific computing
numpy>=1.24.3
scipy>=1.11.1

# Medical imaging
nibabel>=5.1.0
simpleitk>=2.3.1
medpy>=0.4.0

# Data processing
pandas>=2.0.3
scikit-image>=0.21.0
opencv-python>=4.8.0

# Progress bars and utilities
tqdm>=4.65.0
PyYAML>=6.0.1
matplotlib>=3.7.2
seaborn>=0.12.2

# Deep Learning specific
accelerate>=0.20.0
diffusers>=0.19.0

# Medical imaging extensions
nnunetv2>=2.2
batchgenerators>=0.25

# Data downloading
gdown>=4.7.1
synapseclient>=3.0.0

# Optimization and monitoring
wandb>=0.15.0
tensorboard>=2.13.0
omegaconf>=2.3.0

# Image quality metrics
lpips>=0.1.4

# Memory profiling (optional)
memory-profiler>=0.60.0
py3nvml>=0.2.7

# Additional utilities
rich>=13.0.0
colorama>=0.4.6
einops>=0.6.1
timm>=0.9.0

# Configuration management
hydra-core>=1.3.0
typer>=0.9.0

# Development tools
ipython>=8.14.0
jupyter>=1.0.0
jupyterlab>=4.0.5
